fix(volunteer-form): add missing commit and submit voluteer form
