+++ src/app/about/organization-structure/page.tsx
@@ -1,3 +1,4 @@
+import ArticleBody from '#components/ArticleBody';
 import ImageWrapper from '#components/ImageWrapper';
 import Page from '#components/Page';
 import Section from '#components/Section';
@@ -6,9 +7,34 @@ import organizationStructure from '#public/organizational-structure.jpg';
 
 import styles from './page.module.css';
 
+const ceoContent = `## Central Executive Committee
+Nepal Red Cross Society (NRCS) is led by a Central Executive Committee (CEC).
+
+Government of Nepal has formed a 9-member Ad hoc Central Executive Committee (CEC) dated 9 Shrawan 2082. The name list of the committee with designation is as follows:
+
+|                          |                                 |
+|--------------------------|---------------------------------|
+| **Chairman**             | **Members**                     |
+| Mr. Binod Kumar Sharma   | Prof. Dr. Hari Darshan Shrestha |
+| **Vice Chairman**        | Ms. Surya Kumari Shrestha       |
+| Mr. Hari Baral           | Ms. Kanti Rajbhandari           |
+| **Secretary General**    | Dr. Gangadhar Adhikari          |
+| Mr. Min Bahadur Malla    | Mr. Hem Raj Ojha                |
+| **Treasurer General**    |                                 |
+| Mr. Lok Darshan Shrestha |                                 |
+| **Governance Secretariat** |                               |
+| Mr. Sakun Kumar Joshi    |                                 |
+| sakun.joshi@nrcs.org     |                                 |
+`;
+
 export default function AboutUs() {
     return (
         <Page contentClassName={styles.page}>
+            <Section>
+                <ArticleBody
+                    content={ceoContent}
+                />
+            </Section>
             <Section
                 heading="Administrative Structure"
                 childrenContainerClassName={styles.adminStructure}
