+++ src/app/globals.css
@@ -9,7 +9,7 @@
     --color-background-gray: #F3F4F6;
     --color-background-secondary: #f5f5f5;
     --color-foreground: #171717;
-    --color-banner: #024344;
+    --color-banner: #193351;
 
     --color-text-black: #000000;
     --color-text-neutral: #6B7280;
